Created the repositories and controllers + products dtos  and mapper +product service

+++ src/main/java/com/greta/e_shop_api/exposition/dtos/ProductResponseDTO.java
@@ -1,13 +1,11 @@
 package com.greta.e_shop_api.exposition.dtos;
 
-import java.math.BigDecimal;
-
 public record ProductResponseDTO(
         Long id,
         String name,
         String description,
         String imageUrl,
-        BigDecimal price,
+        double price,
         int stock,
         double discount
 ) {}
